...
Brought notation conventions (and American/British spelling) in introduction in line with the rest of the paper.  Minor modifications/additions there too.

Moved "Dijkstra's algorithm" section out of body of paper and into new subsection in introduction.  Scrapped now out of date text that was previously in that section in favour of TODO.  PDF figure is still there, however.

Other minor changes throughout.